universitas: fix add handler
set the delete file will be executed only if the file is not empty

+++ src/Interface/http/api/universitas/handler.ts
@@ -39,7 +39,9 @@ export class UniversitasHandler {
       const data = await this.addUniversitasUsecase.execute(payload);
       res.status(201).json({status: 'success', data});
     } catch (e) {
-      await deleteFile(req.file?.path as string);
+      if (req.file) {
+        await deleteFile(req.file?.path as string);
+      }
       next(e);
     }
   }
